feat: add postpetitiontext content field RE
