+++ src/apis/getPetitionPostsTopLiked.ts
@@ -1,13 +1,17 @@
-import { GetPetitionPostsTopLikedProps } from '@/types/getPetitionTopLiked';
+import { GetBoardPostsProps } from '@/types/getPetitionTopLiked';
 import { client } from './client';
 
-export const getPetitionPostsTopLiked = async ({ page, take }: GetPetitionPostsTopLikedProps) => {
-  const respose = await client.get('/boards/청원게시판/posts/top-liked', {
-    params: {
-      page,
-      take,
-    },
-  });
-
-  return respose.data;
+export const getPetitionPostsTopLiked = async ({ page, take }: GetBoardPostsProps) => {
+  try {
+    const response = await client.get('/boards/청원게시판/posts/top-liked', {
+      params: {
+        page,
+        take,
+      },
+    });
+    console.log(response.data);
+    return response.data;
+  } catch (err) {
+    console.log(err);
+  }
 };
