+++ src/hooks/useGetPetitionPostsTopLiked.ts
@@ -1,9 +1,9 @@
 import { getPetitionPostsTopLiked } from '@/apis/getPetitionPostsTopLiked';
-import { GetPetitionPostsTopLikedProps, PetitionPostsTopLikedResponse } from '@/types/getPetitionTopLiked';
+import { BoardPostsResponse, GetBoardPostsProps } from '@/types/getPetitionTopLiked';
 import { useQuery } from '@tanstack/react-query';
 
-export const useGetPetitionTopLiked = ({ page, take }: GetPetitionPostsTopLikedProps) => {
-  return useQuery<PetitionPostsTopLikedResponse>({
+export const useGetPetitionTopLiked = ({ page, take }: GetBoardPostsProps) => {
+  return useQuery<BoardPostsResponse>({
     queryKey: ['getPetitionTopLiked', page, take],
     queryFn: () => getPetitionPostsTopLiked({ page, take }),
   });
